10-serverless/requirements.txt
keras-image-helper
tflite-runtime
google-cloud-storage